Add a semantic search function
This new panel enables the user to search for functions by meaning. So
for example, when the user enters a query like "performs encryption",
for every function in the program the plugin will send it to the LLM API
along with a prompt to identify whether or not the function performs
encryption. The search results are displayed in a table, and the user
can select the functions in the table to jump to them.

This is a work-in-progress and has a *lot* of issues. A non-exhaustive
list of them:

- Hacks galore, functionality pretty much all handled in one function.
- Race conditions when keeping track of when to change the button text
  back to "Search".
- No progress indicator(s), so it can appear to hang when analyzing
  binaries with lots of functions, even when it's making progress.
- Clicking on an address in the table while a different program is
  selected than the one that was analyzed will not jump to the analyzed
  program.
- No indication of whether the LLM was able to analyze the whole
  function or if it was truncated due to its context window token number
  limit.